Extract reading of instructions into a separate file..
So we can test it separately. Unfortunately, because of the structure
of the main script I needed to place this function to a separate file,
but later it can be put back into one file, after the whole script gets
a testable structure. Better to do this restructuring iteratively. The
problem is that when I import the main asrt script in the test script
the whole experiment is run and so I can't just test one function.

Also tested the file reading code. Some notes about the behavior:
- The algorithm searches for the keywords (e.g. 'inst', 'feedback speed'),
but it's not neccessary to have a full word match. For example 'instagram'
is recognized as 'inst' type instruction.
- The line endings in the text file matters, because they are imported
and used during displaying. See the difference between 'feedback accuracy'
and 'ending' in the default instruction file.

Also some small changes, apart from moveing the function:
- I used platform independent path generation to make this code work
not only on Windows systems.
- Changed the function name a bit to make it clear that a reading is done
by this function.
- Used the 'with' keyword to make sure the file is closed in any case
(python gave me a resource warning about it).