Extract group settings dialog invocation to a separate function
Also fixed a small issue which came up when there was an accent
letter both with a lower and an upper letter in a group name.
For example "Általános". In an unfortunate case this was converted
to "általános" instead of "altalanos". The algorithm first replaced
the "á" letter with "a", then called the lower functions which changed
"Á" to "á" and this lower letter "á" remained in the output.
However this issue is not reproducable consistently, beacuse the
reading order of a dist seems to be random.