Extract computer and display settings dialog to a separate function
Also fix a small issue about the default colors. The original code
appended the default color to the start of the list, which means
that the list contained the default color twice (e.g. as "green"
and also as "Green") which is not a nice gui solution. A better
solution to specify the "initial" attribute of addField,
so the specified item will be selected by default without
adding a duplicate item.