Minimize output: remove derived output variables
Trying to minimize the number of output variables. The idea is to have
a general "core" asrt script, which can be used in lots of asrt
experiments and for this we'll need only the generally usefull,
non-redundant output variables.
The derived ouput variables can be generated anytime after the
experiment script run and the postprocess script can be different
for different experiments. So the "core" script should be the
shared code and it should be as simple and as stable
as possible.