Group all subject data files related code into one class
Additional changes:
- For the output txt file we check whether the file
exists to find out that we need the header line or not.
- participants_in_experiment.txt is updated at the same
time as we save the new subject ID in "participant" shelve file.
These two files contains the same data in different forms.
- Removed the code which tried to write out "asrt_type" and "PCode"
in to the participants_in_experiment.txt file since these attributes
has no meaning (there might be more asrt types for the same subject in
different sessions).

The long list of input paramters or retuned variabled will
be simplified later.