Introduce Experiment class
Move all variables into this class which are used in more
methods during the experiment (and is not part of other classes).
So we don't need to pass a lot of parameters to functions
or return with a lot of values. Having a class is better,
than having these variables global, because we can test the code
easier if we can control the lifetime of these variables.